requirements.txt
streamlit==1.50.0
opencv-python==4.12.0.88
numpy==2.2.6
pillow==11.3.0
